GET-requied shcema
- writing get-req schema
- take out chemas in separate files
