+++ app/SBooking.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from datetime import date
+
+class SBooking(BaseModel):
+    room_id: int
+    date_from: date
+    date_to: date
